requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
qdrant-client
google-generativeai
pypdf
jinja2
python-multipart
